requirements.txt
python-dotenv==0.21.1
discord.py>=2.1.1